pyproject.toml
[project]
name = "whiteanalysis"
version = "0.1.0"
description = "An analysis of HC White's work."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instructor>=1.6.3",
    "isort>=5.13.2",
    "openai>=1.54.4",
    "pdfminer>=20191125",
    "pre-commit>=4.0.1",
    "pydantic>=2.9.2",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.3",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
    "tqdm>=4.67.0",
    "typer>=0.13.0",
    "unstructured[pdf] >= 0.16.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
whiteanalysis = "whiteanalysis.main:run"
